+++ src/handler/handlerUpload.js
@@ -1,4 +1,3 @@
-//import path from 'node:path';
 import multer from 'multer';
 
 /**
@@ -12,21 +11,14 @@ import multer from 'multer';
 
 /**
  * Create the upload middleware.
+ * @param {number} [uploadFileSizeLimit] - Maximum size of each uploaded file in bytes or no limit if omitted.
  * @returns {RequestHandler} - Request handler.
  */
-export const handlerUpload = () => {
+export const handlerUpload = (uploadFileSizeLimit = Infinity) => {
 	const upload = multer({
-		storage: multer.diskStorage({
-			/*
-			destination: '/tmp/webplsql',
-			filename: (req, file, cb) => {
-				const uniqueSuffix = `${Date.now()}-${Math.round(Math.random() * 1e9)}`;
-				cb(null, `${file.fieldname}-${uniqueSuffix}${path.extname(file.originalname)}`);
-			},
-			*/
-		}),
+		storage: multer.diskStorage({}),
 		limits: {
-			fileSize: 50 * 1024 * 1024, // 50MB limit
+			fileSize: uploadFileSizeLimit,
 		},
 	});